% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_ul.R
\name{scale_fill_ul}
\alias{scale_fill_ul}
\title{Fill scale for University of Ljubljana color palettes}
\usage{
scale_fill_ul(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  midpoint = NA,
  ...
)
}
\arguments{
\item{palette}{Character. The name of the palette to use.
- Main: \code{"main"} (default)
- Secondary (cold): \code{"secondaryCold"}
- Secondary (warm): \code{"secondaryWarm"}
- Individual colors: \code{"red"}, \code{"antracit"}, \code{"medium"}, \code{"lajt"}, \code{"darkblue"},
\code{"navyblue"}, \code{"turquoise"}, \code{"green"}, \code{"yellow"}, \code{"orange"}, \code{"burgundy"},
\code{"pink"}
- Divergent: \code{"divergent1"}, \code{"divergent2"}, \code{"divergent3"}, \code{"divergent4"}}

\item{discrete}{Logical. If \code{TRUE} (default), a discrete color scale is used. If \code{FALSE}, a continuous color gradient is used.}

\item{reverse}{Logical. If \code{TRUE}, reverses the order of colors in the palette. Default is \code{FALSE}.}

\item{midpoint}{The midpoint (in data value) of the diverging scale.}

\item{...}{Additional arguments to be passed to the discrete_scale or scale_fill_gradientn functions.}
}
\value{
A fill scale that can be added to a ggplot2 plot.
}
\description{
This function provides fill scales using the official visual identity
color palettes of the University of Ljubljana for use in ggplot2.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(class, hwy)) +
  geom_boxplot(aes(fill = class)) +
  scale_fill_ul("primary", discrete = TRUE)
}

}
\seealso{
\code{\link[ggplot2]{scale_fill_discrete}}, \code{\link[ggplot2]{scale_fill_gradientn}}
}
