% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ul.R
\name{theme_ul}
\alias{theme_ul}
\title{Theme UL}
\usage{
theme_ul(
  legend.position = "top",
  panel.background.fill = "#E8E9EA",
  plot.background.fill = "white",
  panel.grid.major.color = NULL,
  legend.justification = c(0, 1),
  legend.key = element_rect(fill = "transparent"),
  ...
)
}
\arguments{
\item{legend.position}{A character string that specifies the position of the legend. Default is \code{"top"}.}

\item{panel.background.fill}{A character string that specifies the HEX code for the color of the panel background. Default is \code{"#E8E9EA"}.}

\item{plot.background.fill}{A character string that specifies the HEX code for the color of the plot background. Default is \code{"white"}.}

\item{panel.grid.major.color}{A character string that specifies the HEX code for the color of the major grid lines. If left as \code{NULL} (the default setting), the color will be automatically set according to the \code{panel.background.fill} parameter.}

\item{legend.justification}{A numeric vector of length 2 that determines the justification of the legend. Default is \code{c(0, 1)}.}

\item{legend.key}{An \code{element_rect} object defining the appearance of the legend key. Default is \code{element_rect(fill = "transparent")}.}

\item{...}{Other arguments passed on to the function `theme()`.}
}
\value{
A ggplot2 theme object that can be added to a ggplot.
}
\description{
This function creates a ggplot2 theme that follows the visual identity of the University of Ljubljana.
}
\examples{
\dontrun{
ggplot(data = mtcars, aes(x = disp, y = mpg, color = qsec, size = wt)) +
geom_point() +
labs(y = "Miles per gallon",
x = "Engine size (cu. in.) ",
title = "Larger engine consume more gas",
subtitle = "V-shape engines are typically larger",
caption = "Data source: Motor Trend US magazine.",
color = "1/4 mile time (seconds)",
size = "Weight (1000 lbs)") +
scale_color_ul(palette = "red", discrete = FALSE) +
theme_ul(plot.background.fill = ul_color("lajt"))
}
}
\seealso{
\code{\link[ggplot2]{theme}}
}
