% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul_color.R
\name{ul_color}
\alias{ul_color}
\title{UL color theme}
\usage{
ul_color(...)
}
\arguments{
\item{...}{Optional character vectors. If color names are supplied, only those colors
will be returned. If no arguments are supplied, all available colors are returned.}
}
\value{
A character vector of hexadecimal color codes.
If no arguments are supplied, all color codes are returned.
If specific color names are supplied, only the color codes for those colors are returned.
}
\description{
This function provides a set of color schemes adhering to the official visual identity
of the University of Ljubljana. The colors are categorized into primary, secondary cold,
and secondary warm color schemes. If no arguments are supplied, it returns all
available colors. If specific color names are supplied, it will return only
those colors.
}
\details{
The primary color scheme consists of red, antracit, medium, and lajt.
The secondary cold color scheme includes darkblue, navyblue, turquoise, and green.
The secondary warm color scheme is defined by yellow, orange, burgundy, and pink.
}
\examples{
ul_color("red", "green") # Returns hexadecimal color codes for red and green
ul_color() # Returns all available colors

}
