% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul_pal.R
\name{ul_pal}
\alias{ul_pal}
\title{University of Ljubljana color palettes}
\usage{
ul_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character. The name of the palette to use.
- Main: \code{"main"} (default)
- Secondary (cold): \code{"secondaryCold"}
- Secondary (warm): \code{"secondaryWarm"}
- Individual colors: \code{"red"}, \code{"antracit"}, \code{"medium"}, \code{"lajt"}, \code{"darkblue"},
\code{"navyblue"}, \code{"turquoise"}, \code{"green"}, \code{"yellow"}, \code{"orange"}, \code{"burgundy"},
\code{"pink"}
- Divergent: \code{"divergent1"}, \code{"divergent2"}, \code{"divergent3"}, \code{"divergent4"}}

\item{reverse}{Logical. If \code{TRUE}, reverses the order of colors in the palette. Default is \code{FALSE}.}

\item{...}{Additional arguments to be passed to the \code{colorRampPalette} function.}
}
\value{
A function that takes an integer argument n and returns a character vector of n colors.
}
\description{
This function provides color palettes adhering to the official visual identity
of the University of Ljubljana. It includes the main palette, primary and secondary color schemes,
individual color gradations, and multiple divergent color palettes.
}
\examples{
\dontrun{
pal <- ul_pal("primary")
pal(5)  # Generates a palette of 5 colors based on the "primary" palette

pal <- ul_pal("divergent1", reverse = TRUE)
pal(7)  # Generates a reversed divergent palette of 7 colors
}
}
