% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ul_color.R
\name{ul_color}
\alias{ul_color}
\title{University of Ljubljana colors}
\usage{
ul_color(...)
}
\arguments{
\item{...}{Optional character vectors. If color names are supplied, only those colors
will be returned. If no arguments are supplied, all available colors are returned. Avaiable colors are
red, antracit, medium, lajt, darkblue, navyblue, turquoise, green, yellow, orange, burgundy, and pink.}
}
\value{
A character vector of hexadecimal color codes.
If no arguments are supplied, all color codes are returned.
If specific color names are supplied, only the color codes for those colors are returned.
}
\description{
This function provides a set of color schemes adhering to the official visual identity
of the University of Ljubljana. If no arguments are supplied, it returns all
available colors. If specific color names are supplied, it will return only
those colors.
}
\examples{
ul_color("red", "green") # Returns hexadecimal color codes for red and green
ul_color() # Returns all available colors

}
